[Experimental. Please report on the forum your opinion regarding these changes] http://asf.mangaheart.org/index.php?menu=5&f=2&t=113
Add:
The file explorer is now activated on a case-by-case basis when downloading (whether a filter is found or not).
It's the first step in prevision of ASF2.0 filter-by-filter basis.

Change:
For better clarity, the option to select whether the file explorer should be displayed is now written like in Firefox's options.

Fix:
Fixed the suggested folder's display if the saved folder is forced to Desktop or Download.


Please report any problem you encounter with this revision.
If it doesn't work correctly, use r115 instead and let me know on the forum.